services/chat/src/api/requirements.txt
# General
python-dotenv==1.0.1
jsonlines
pandas
requests

# API server
fastapi
uvicorn[standard]

# Google Cloud Services
google-cloud-aiplatform
google-cloud-discoveryengine
google-auth
vertexai

# OpenTelemetry and tracing
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi

# Prompty (simplified)
prompty

# Database
prisma

# Local LLM & Search
litellm
chromadb
sentence-transformers
